// src/app/pages/transfer/view-details-transfer/view-details-rates/shared/view-details-rates-enum.ts
export enum ViewDetailsRatesEnum {
    title = 'Rates and Prices',
    idTitle = 'Id',
    vehicleTitle = 'Vehicle',
    originTitle = 'Origin',
    destinationTitle = 'Destination',
    serviceModeTitle = 'Service Mode',
    peopleTitle = 'People',
    taxesTitle = 'Taxes',
    extraChargesTitle = 'Extra Charges',
    parkingPriceTitle = 'Parking Price',
    totalPriceTitle = 'Total Price',

    id = 'id',
    vehicle = 'vehicle',
    origin = 'origin',
    destination = 'destination',
    serviceMode = 'serviceMode',
    people = 'people',
    taxes = 'taxes',
    extraCharges = 'extraCharges',
    parkingPrice = 'parkingPrice',
    totalPrice = 'totalPrice',

    selectedTransferIdCookieName = 'selectedTransferId',
    buttonRates = 'Setup Rates and Prices',
    yes = 'Yes',
    no = 'No',
}